Remove USING TIMESTAMP clause from data points insert statement.
YugaByte is strongly consistent, and update order is controlled by the RAFT replication sequence.
Therefore, when USING TIMESTAMP is used, YugaByte DB would incur a read penalty to make sure that
an insert with a lower timestamp does not actually take place which is unnecessary.

Under the covers, YugaByte DB implements user specified timestamps (i.e. the USING TIMESTAMP clause)
as an additional implicit column, which needs to be read on each insert.